Estimate confidence intervals for binomial distributions

This function is an almost exact copy of the asbio::ci.p() function. The only differences are that (1) the plot argument and associated functionality have been removed, and (2) the output head has been slighlty altered.

ci_binomial(
  data,
  conf = 0.95,
  summarized = FALSE,
  phat = NULL,
  fpc = FALSE,
  n = NULL,
  N = NULL,
  method = "agresti.coull"
)

Arguments

data

A vector of binary data. Required if summarized = FALSE.

conf

Level of confidence 1 - P(type I error).

summarized

Logical; indicate whether raw data or summary stats are to be used.

phat

Estimate of π. Required if summarized = TRUE.

fpc

Logical. Indicates whether finite population corrections should be used. If fpc = TRUE then N must be specified. Finite population corrections are not possible for method = "exact" or method = "score".

n

Sample size. Required if summarized = TRUE.

N

Population size. Required if fpc = TRUE.

method

Type of method to be used in confidence interval calculations, method ="agresti.coull" is the default. Other procedures include method = "asymptotic" which provides the conventional normal (Wald) approximation, method = "score", method = "LR", and method = "exact" (see Details below). Partial names can be used. The "exact" method cannot be implemented if summarized=TRUE.

Value

Returns a list of class = "ci".

Details

See ?asbio::ci.p for details.

Author

See ?asbio::ci.p