Specify projections for multiple columns

The transition from a metadata-rich dataframe consisting of projectable_cols to an ordinary dataframe that is fit for presentation is guided by the shadow attributes attached to each column of the initial dataframe. The prj_shadow_*() functions are designed to make it easy to specify those shadow attributes for multiple columns at once.

prj_shadow_all(.data, .shadow)

prj_shadow_if(.data, .predicate, .shadow)

prj_shadow_at(.data, .vars, .shadow)

Arguments

.data
A dataframe, ideally one containing projectabel_cols

.shadow
A character vector containing glue-like specifications to which to set the shadow attribute of the relevant columns.

.predicate
A predicate function to be applied to the columns or a logical vector. The variables for which .predicate is or returns TRUE have their shadow updated

.vars
A character vector containing the names of columns in .data of which to update the shadow attribute

Value

A dataframe, the columns of which have updated shadow attributes.

Examples

# Create a table made up of `projectable_col`s
my_tbl <- prj_tbl_rows(
  .data = mtcars,
  Cylinders = cyl,
  Transmission = list(Automatic = am %in% 0, Manual = am %in% 1),
)

my_tbl <- prj_tbl_cols(
  .data = my_tbl,
  `V-Shaped` = col_freq(n = vs %in% 1, N = vs %in% 0:1),
  `Not V-shaped` = col_freq(n = vs %in% 0, N = vs %in% 0:1)
)

my_tbl <- prj_tbl_summarise(.data = my_tbl)

# Update the `shadow` attributes
my_tbl <- prj_shadow_if(my_tbl, is_col_freq(.), c(Count = "{n}", Proportion = "{signif(p, 2)}"))

# Project it back into an ordinary dataframe
prj_project(my_tbl)

#> # A tibble: 5 x 6
#>   row_spanner rows  `V-Shaped.Count` `V-Shaped.Propo… `Not V-shaped.C…
#> * <chr>       <chr> <glue>           <glue>           <glue>          
#> 1 Cylinders   4     10               0.91             1               
#> 2 Cylinders   6     4                0.57             3               
#> 3 Cylinders   8     0                0                14              
#> 4 Transmissi… Auto… 7                0.37             12              
#> 5 Transmissi… Manu… 7                0.54             6               
#> # … with 1 more variable: `Not V-shaped.Proportion` <glue>