Project metadata-rich tables

The prj_project() function is designed to take in a dataframe made up of projectable_cols, and to 'project' it into an ordinary dataframe as per the instructions provided in the shadow attribute of each column. The prj_gt() function does the same thing, but initialises a gt object for display.

prj_project(.data, .cols = list())

prj_gt(
  .data,
  .cols = list(),
  rowgroup_col = "row_spanner",
  rowname_col = "rows",
  ...
)

Arguments

.data
A dataframe, ideally one containing projectable_cols

.cols
A named list. Each name should be the name of a column in .data; each value should be a named character vector containing glue-like specifications for the output columns.

rowgroup_col
The name of a column in .data to group rows by; if NULL no grouping will be used.

rowname_col
The name of a column in .data to take as the row labels; if NULL no row labels will be applied

...
Additional arguments to pass on to gt::gt()

Value

a projection object

Details

The shadow attribute of each column can be set via the .cols argument of prj_project() or by using the prj_shadow_if() and prj_shadow_at() helper functions.

The projection output will also come attached with metadata which keeps track of which columns in the output belong to which columns in the input.

Examples

# Create a table made up of `projectable_col`s
my_tbl <- prj_tbl_rows(
  .data = mtcars,
  Cylinders = cyl,
  Transmission = list(Automatic = am %in% 0, Manual = am %in% 1),
)

my_tbl <- prj_tbl_cols(
  .data = my_tbl,
  `V-Shaped` = col_freq(n = vs %in% 1, N = vs %in% 0:1),
  `Not V-shaped` = col_freq(n = vs %in% 0, N = vs %in% 0:1)
)

my_tbl <- prj_tbl_summarise(.data = my_tbl)

# Project it back into an ordinary dataframe
prj_project(my_tbl, list(
  `V-Shaped` = "{signif(p, 2)} ({n})",
  `Not V-shaped` = "{signif(p, 2)} ({n})"
))

#> # A tibble: 5 x 4
#>   row_spanner  rows      `V-Shaped` `Not V-shaped`
#> * <chr>        <chr>     <glue>     <glue>        
#> 1 Cylinders    4         0.91 (10)  0.091 (1)     
#> 2 Cylinders    6         0.57 (4)   0.43 (3)      
#> 3 Cylinders    8         0 (0)      1 (14)        
#> 4 Transmission Automatic 0.37 (7)   0.63 (12)     
#> 5 Transmission Manual    0.54 (7)   0.46 (6)      

# Produce a `gt` display object
prj_gt(my_tbl, list(
  `V-Shaped` = "{signif(p, 2)} ({n})",
  `Not V-shaped` = "{signif(p, 2)} ({n})"
))

#> <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #xqpnnehfrc .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #xqpnnehfrc .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #xqpnnehfrc .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #xqpnnehfrc .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 0;
#>   padding-bottom: 4px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #xqpnnehfrc .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #xqpnnehfrc .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #xqpnnehfrc .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #xqpnnehfrc .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #xqpnnehfrc .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #xqpnnehfrc .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #xqpnnehfrc .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #xqpnnehfrc .gt_group_heading {
#>   padding: 8px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #xqpnnehfrc .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #xqpnnehfrc .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #xqpnnehfrc .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #xqpnnehfrc .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #xqpnnehfrc .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 12px;
#> }
#> 
#> #xqpnnehfrc .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #xqpnnehfrc .gt_first_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #xqpnnehfrc .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #xqpnnehfrc .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #xqpnnehfrc .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #xqpnnehfrc .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #xqpnnehfrc .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #xqpnnehfrc .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #xqpnnehfrc .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #xqpnnehfrc .gt_sourcenote {
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #xqpnnehfrc .gt_left {
#>   text-align: left;
#> }
#> 
#> #xqpnnehfrc .gt_center {
#>   text-align: center;
#> }
#> 
#> #xqpnnehfrc .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #xqpnnehfrc .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #xqpnnehfrc .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #xqpnnehfrc .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #xqpnnehfrc .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #xqpnnehfrc .gt_footnote_marks {
#>   font-style: italic;
#>   font-size: 65%;
#> }
#> </style>
#> <div id="xqpnnehfrc" style="overflow-x:auto;overflow-y:auto;width:auto;height:auto;"><table class="gt_table">
#>   
#>   <thead class="gt_col_headings">
#>     <tr>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1"></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1">V-Shaped</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1">Not V-shaped</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr class="gt_group_heading_row">
#>       <td colspan="3" class="gt_group_heading">Cylinders</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left gt_stub">4</td>
#>       <td class="gt_row gt_center">0.91 (10)</td>
#>       <td class="gt_row gt_center">0.091 (1)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left gt_stub">6</td>
#>       <td class="gt_row gt_center">0.57 (4)</td>
#>       <td class="gt_row gt_center">0.43 (3)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left gt_stub">8</td>
#>       <td class="gt_row gt_center">0 (0)</td>
#>       <td class="gt_row gt_center">1 (14)</td>
#>     </tr>
#>     <tr class="gt_group_heading_row">
#>       <td colspan="3" class="gt_group_heading">Transmission</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left gt_stub">Automatic</td>
#>       <td class="gt_row gt_center">0.37 (7)</td>
#>       <td class="gt_row gt_center">0.63 (12)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left gt_stub">Manual</td>
#>       <td class="gt_row gt_center">0.54 (7)</td>
#>       <td class="gt_row gt_center">0.46 (6)</td>
#>     </tr>
#>   </tbody>
#>   
#>   
#> </table></div>